r flytter egenvurdering ut av innhold og inn i situasjonsbestemt komponenten
